refactor inheritance
