+++ magpylib/_src/obj_classes/class_current_Line.py
@@ -3,10 +3,7 @@ DOCSTRINGS V4 READY
 """
 from magpylib._src.fields.field_BH_line import current_vertices_field
 from magpylib._src.input_checks import check_format_input_vertices
-from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseExcitations import BaseCurrent
-from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
-from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
 from magpylib._src.utility import Registered
 
 
@@ -21,7 +18,7 @@ from magpylib._src.utility import Registered
         "segment_end": 2,
     },
 )
-class Line(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseCurrent):
+class Line(BaseCurrent):
     """Current flowing in straight lines from vertex to vertex.
 
     Can be used as `sources` input for magnetic field computation.
@@ -113,9 +110,7 @@ class Line(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseCurrent):
         self.vertices = vertices
 
         # init inheritance
-        BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
-        BaseDisplayRepr.__init__(self)
-        BaseCurrent.__init__(self, current)
+        super().__init__(position, orientation, current, style, **kwargs)
 
     # property getters and setters
     @property
