+++ magpylib/_src/obj_classes/class_current_Loop.py
@@ -3,10 +3,7 @@ DOCSTRINGS V4 READY
 """
 from magpylib._src.fields.field_BH_loop import current_loop_field
 from magpylib._src.input_checks import check_format_input_scalar
-from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
 from magpylib._src.obj_classes.class_BaseExcitations import BaseCurrent
-from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
-from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
 from magpylib._src.utility import Registered
 
 
@@ -16,7 +13,7 @@ from magpylib._src.utility import Registered
     field_func=current_loop_field,
     source_kwargs_ndim={"current": 1, "diameter": 1},
 )
-class Loop(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseCurrent):
+class Loop(BaseCurrent):
     """Circular current loop.
 
     Can be used as `sources` input for magnetic field computation.
@@ -102,9 +99,7 @@ class Loop(BaseGeo, BaseDisplayRepr, BaseGetBH, BaseCurrent):
         self.diameter = diameter
 
         # init inheritance
-        BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
-        BaseDisplayRepr.__init__(self)
-        BaseCurrent.__init__(self, current)
+        super().__init__(position, orientation, current, style, **kwargs)
 
     # property getters and setters
     @property
