+++ magpylib/_src/obj_classes/class_misc_Dipole.py
@@ -3,9 +3,7 @@ DOCSTRINGS V4 READY
 """
 from magpylib._src.fields.field_BH_dipole import dipole_field
 from magpylib._src.input_checks import check_format_input_vector
-from magpylib._src.obj_classes.class_BaseDisplayRepr import BaseDisplayRepr
-from magpylib._src.obj_classes.class_BaseGeo import BaseGeo
-from magpylib._src.obj_classes.class_BaseGetBH import BaseGetBH
+from magpylib._src.obj_classes.class_BaseExcitations import BaseSource
 from magpylib._src.utility import Registered
 
 
@@ -15,7 +13,7 @@ from magpylib._src.utility import Registered
     field_func=dipole_field,
     source_kwargs_ndim={"moment": 2},
 )
-class Dipole(BaseGeo, BaseDisplayRepr, BaseGetBH):
+class Dipole(BaseSource):
     """Magnetic dipole moment.
 
     Can be used as `sources` input for magnetic field computation.
@@ -97,8 +95,7 @@ class Dipole(BaseGeo, BaseDisplayRepr, BaseGetBH):
         self.moment = moment
 
         # init inheritance
-        BaseGeo.__init__(self, position, orientation, style=style, **kwargs)
-        BaseDisplayRepr.__init__(self)
+        super().__init__(position, orientation, style, **kwargs)
 
     # property getters and setters
     @property
